+++ Engine/Templates/2D/main.cpp
@@ -20,5 +20,4 @@ int main() {
     }
   );
   game.Run();
-  return 0;
 }